Conflicts:
	JFSService/src/main/java/jfs/service/services/JobOfferService.java

+++ JFSData/src/main/java/jfs/data/stores/JobOfferStore.java
@@ -14,10 +14,14 @@ import org.bson.Document;
 import org.bson.types.ObjectId;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
  * Created by lpuddu on 12-11-2015.
+ *
+ * Class used for access to the job offer store
+ *
  */
 public class JobOfferStore extends DataStore {
     public static final JobOfferStore store = new JobOfferStore();
@@ -27,6 +31,7 @@ public class JobOfferStore extends DataStore {
         super("joboffers");
     }
 
+    //Get all job offers
     public List<JobOfferDO> getAllOffers(){
         List<JobOfferDO> offers = new ArrayList<JobOfferDO>();
         FindIterable<DBObject> results = this.collection.find();
@@ -45,6 +50,7 @@ public class JobOfferStore extends DataStore {
         return offers;
     }
 
+    //Add a new job offer by JobOfferDO and companyId
     public Boolean addOffer(JobOfferDO offer, String companyId){
         if (offer != null) {
             return this.insert(offer) && this.metricsStore.add(offer._id, companyId);
@@ -53,6 +59,7 @@ public class JobOfferStore extends DataStore {
         }
     }
 
+    //Add a several job offer by List<JobOfferDO> and companyId
     public Boolean addOffers(List<JobOfferDO> offers, String companyId){
         List<DBObject> docs = this.createDocumentList(offers);
         try{
@@ -65,6 +72,7 @@ public class JobOfferStore extends DataStore {
         }
     }
 
+    //Get a specific job offer by id
     public JobOfferDO getById(String id){
         DBObject obj = (DBObject)this.collection.find(new BasicDBObject("_id", id)).first();
         if(obj != null){
@@ -74,10 +82,12 @@ public class JobOfferStore extends DataStore {
         }
     }
 
+    //Get a list of the job offers by a search term, limited to 100 findings
     public List<JobOfferDO> getJobOffersText(String term){
         return this.getJobOffersText(term, 100);
     }
 
+    //Get a list of the job offers by a search term and an amount
     public List<JobOfferDO> getJobOffersText(String term, int amount){
         List<JobOfferDO> offers = new ArrayList<JobOfferDO>();
         FindIterable<DBObject> results = this.collection.find(new BasicDBObject("textSearch", term)).limit(amount);
@@ -87,6 +97,7 @@ public class JobOfferStore extends DataStore {
         return offers;
     }
 
+    //Get a list of the job offers sorted by recency and a maximum amount
     public List<JobOfferDO> getJobOffersRecent(int amount){
         List<JobOfferDO> offers = new ArrayList<JobOfferDO>();
         FindIterable<DBObject> results = this.collection.find().sort(Sorts.descending("_id")).limit(amount);
@@ -96,6 +107,9 @@ public class JobOfferStore extends DataStore {
         return offers;
     }
 
+    //Get a list of the job offers by a specific studentSubscriptionDO and jobOfferVisibilityBuffer
+    //jobOfferVisibilityBuffer is giving as a long and represents the seconds since the job offer was created
+    //all offers that were created before the seconds will not be considered even if they would quality the criteria in the jobOfferVisibilityBuffer
     public List<JobOfferDO> getJobOffersByCriteria(StudentSubscriptionsDO studentSubscriptionsDO, long jobOfferVisibilityBuffer){
 
         ArrayList<Pair<String, Object>> pairs = new ArrayList<Pair<String, Object>>();
@@ -114,9 +128,26 @@ public class JobOfferStore extends DataStore {
             pairs.add(new Pair("location", new BasicDBObject("$regex", studentSubscriptionsDO.location)));
         }
         if(studentSubscriptionsDO.skills != null && !"".equals(studentSubscriptionsDO.skills)){
-            //List<String> Skills = Arrays.asList(studentSubscriptionsDO.skills.split("\\s*,\\s*")); //also deleted any additional white spaces
-            String skillsRegex = studentSubscriptionsDO.skills.replace(",", "|");
-            skillsRegex = skillsRegex.replace(" ", "");
+            /*
+                For easier creation of the search string and array is created
+                which also removes all whitespaces before or after commas that are seperating the skills
+                \\Q...\\E is used to quote string in Regex. This is necessary in order to not take e.g. the + character as a command
+                | is used to combine all skills by OR
+            */
+            List<String> myList = Arrays.asList(studentSubscriptionsDO.skills.split("\\s*,\\s*"));
+
+            String skillsRegex = "\\Q";
+            if (myList.size() > 1) {
+                for (int i = 0; i < myList.size(); i++){
+                    skillsRegex += myList.get(i);
+                    if (i != myList.size() -1) {
+                        skillsRegex += "\\E|\\Q";
+                    }
+                }
+            }else{
+                skillsRegex += myList.get(0);
+            }
+            skillsRegex += "\\E";
 
             pairs.add(new Pair("skills", new BasicDBObject("$regex", skillsRegex)));
         }
@@ -125,6 +156,8 @@ public class JobOfferStore extends DataStore {
         return doList;
     }
 
+    //Function to search for specific job offers by pairs of search terms
+    //Reused in other functions
     public List<JobOfferDO> getJobOffers(List<Pair<String, Object>> pairs){
         List<JobOfferDO> offers = new ArrayList<JobOfferDO>();
         Document query = new Document();
@@ -139,17 +172,27 @@ public class JobOfferStore extends DataStore {
     }
 
     public boolean delete(String jobOfferId){
-        BasicDBObject filter = new BasicDBObject("_id", jobOfferId);
-        DeleteResult result = this.collection.deleteOne(filter);
-        long count = result.getDeletedCount();
-        if (count == 1){
-            return true;
-        }
-        else{
-            return false;
+        if (jobOfferId != null || !jobOfferId.isEmpty()) {
+            BasicDBObject filter = new BasicDBObject("_id", jobOfferId);
+            DeleteResult result = this.collection.deleteOne(filter);
+            return result.wasAcknowledged();
+        }else{
+            throw new NullPointerException("jobOfferId parameter is null");
         }
     }
 
+    public boolean deleteJobOffers(String companyId){
+        if (companyId != null || !companyId.isEmpty()) {
+            BasicDBObject filter = new BasicDBObject("userId", companyId);
+            DeleteResult result = this.collection.deleteMany(filter);
+            return result.wasAcknowledged();
+        }else{
+            throw new NullPointerException("companyId parameter is null or empty");
+        }
+    }
+
+    //deserialize a JobOfferDO from a DBObject
+    //Reused in other functions
     private JobOfferDO extractJobOffer(DBObject object){
         return this.serializer.deSerialize(object.toString(), JobOfferDO.class);
     }
